+++ src/shogun/io/StreamingAsciiFile.cpp
@@ -150,7 +150,7 @@ GET_FLOAT_VECTOR(float64_t)
 /* Methods for reading a dense vector and a label from an ascii file */
 
 #define GET_VECTOR_AND_LABEL(fname, conv, sg_type)			\
-	void CStreamingAsciiFile::get_vector_and_label(sg_type*& vector, int32_t& num_feat, float32_t& label) \
+	void CStreamingAsciiFile::get_vector_and_label(sg_type*& vector, int32_t& num_feat, float64_t& label) \
 	{								\
 		char* buffer = NULL;					\
 		ssize_t bytes_read;					\
@@ -233,7 +233,7 @@ GET_VECTOR_AND_LABEL(get_longreal_vector_and_label, atoi, floatmax_t)
 #undef GET_VECTOR_AND_LABEL
 
 #define GET_FLOAT_VECTOR_AND_LABEL(sg_type)				\
-	void CStreamingAsciiFile::get_vector_and_label(sg_type*& vector, int32_t& len, float32_t& label) \
+	void CStreamingAsciiFile::get_vector_and_label(sg_type*& vector, int32_t& len, float64_t& label) \
 	{								\
 		char *line=NULL;					\
 		int32_t num_chars = buf->read_line(line);		\
@@ -315,7 +315,7 @@ GET_STRING(get_longreal_string, atoi, floatmax_t)
 /* Methods for reading a string vector and a label from an ascii file */
 
 #define GET_STRING_AND_LABEL(fname, conv, sg_type)			\
-void CStreamingAsciiFile::get_string_and_label(sg_type*& vector, int32_t& len, float32_t& label) \
+void CStreamingAsciiFile::get_string_and_label(sg_type*& vector, int32_t& len, float64_t& label) \
 {									\
 	char* buffer = NULL;						\
 	ssize_t bytes_read;						\
@@ -469,7 +469,7 @@ GET_SPARSE_VECTOR(get_longreal_sparse_vector, atoi, floatmax_t)
 /* Methods for reading a sparse vector and a label from an ascii file */
 
 #define GET_SPARSE_VECTOR_AND_LABEL(fname, conv, sg_type)			\
-void CStreamingAsciiFile::get_sparse_vector_and_label(SGSparseVectorEntry<sg_type>*& vector, int32_t& len, float32_t& label) \
+void CStreamingAsciiFile::get_sparse_vector_and_label(SGSparseVectorEntry<sg_type>*& vector, int32_t& len, float64_t& label) \
 {									\
 	char* buffer = NULL;						\
 	ssize_t bytes_read;						\
